refactor: App 의 도메인 관련된 메소드 Util 메소드로 이동
